requirements.txt
Flask==2.3.3
pycryptodome
requests